# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BookingMode": {
        "enum": [
          "internal_booking",
          "external_booking_link"
        ],
        "title": "BookingMode",
        "type": "string"
      },
...
      "BusinessCreate": {
        "properties": {
          "booking_mode": {
            "$ref": "#/components/schemas/BookingMode",
            "default": "internal_booking"
          },
          "external_booking_label": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Booking Label"
          },
          "external_booking_provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExternalBookingProvider"
              },
              {
                "type": "null"
              }
            ]
          },
          "external_booking_url": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Booking Url"
          },
          "language": {
            "$ref": "#/components/schemas/BusinessLanguage",
            "default": "en"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "subscription_plan": {
            "$ref": "#/components/schemas/SubscriptionPlan",
            "default": "full_booking"
          },
          "timezone": {
            "examples": [
              "Europe/Warsaw"
            ],
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          },
          "transfer_destination_policy": {
            "$ref": "#/components/schemas/TransferDestinationPolicy",
            "default": "business_phone"
          },
          "transfer_enabled": {
            "default": false,
            "title": "Transfer Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "timezone"
        ],
        "title": "BusinessCreate",
        "type": "object"
      },
...
      "BusinessRead": {
        "properties": {
          "booking_mode": {
            "title": "Booking Mode",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "external_booking_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Booking Label"
          },
          "external_booking_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Booking Provider"
          },
          "external_booking_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Booking Url"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "subscription_plan": {
            "title": "Subscription Plan",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "transfer_destination_policy": {
            "title": "Transfer Destination Policy",
            "type": "string"
          },
          "transfer_enabled": {
            "title": "Transfer Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "timezone",
          "language",
          "phone",
          "is_active",
          "transfer_enabled",
          "transfer_destination_policy",
          "booking_mode",
          "external_booking_url",
          "external_booking_label",
          "external_booking_provider",
          "subscription_plan",
          "created_at"
        ],
        "title": "BusinessRead",
        "type": "object"
      },
...
      "BusinessUpdate": {
        "properties": {
          "booking_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookingMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "external_booking_label": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Booking Label"
          },
          "external_booking_provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExternalBookingProvider"
              },
              {
                "type": "null"
              }
            ]
          },
          "external_booking_url": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Booking Url"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BusinessLanguage"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "subscription_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubscriptionPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "transfer_destination_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransferDestinationPolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "transfer_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer Enabled"
          }
        },
        "title": "BusinessUpdate",
        "type": "object"
      },
      "ClientCreate": {
        "properties": {
          "customer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "name"
        ],
        "title": "ClientCreate",
        "type": "object"
      },
      "ClientRead": {
        "properties": {
          "business_id": {
            "title": "Business Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "business_id",
          "customer_id",
          "name",
          "email",
          "phone",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "ClientRead",
        "type": "object"
      },
      "ClientUpdate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "title": "ClientUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "bearerFormat": "JWT",
        "description": "Use the access token returned by login or refresh. Send it as `Authorization: Bearer <access_token>`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Appointment Voice SaaS API \u2014 multi-tenant booking and availability backend built on a production FastAPI foundation (JWT auth, admin workflows, audit logs, Redis rate limiting, object storage). Salon domain: businesses, staff, services, hours, availability slots, bookings with DB-level double-booking protection. IVR/SMS/calendar: planned.",
    "summary": "Multi-tenant booking and availability backend for local service businesses",
    "title": "Appointment Voice SaaS API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/businesses": {
      "get": {
        "operationId": "list_businesses_endpoint_api_v1_businesses_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_inactive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Inactive",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BusinessRead"
                  },
                  "title": "Response List Businesses Endpoint Api V1 Businesses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Businesses Endpoint",
        "tags": [
          "businesses"
        ]
      },
      "post": {
        "operationId": "create_business_endpoint_api_v1_businesses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Business Endpoint",
        "tags": [
          "businesses"
        ]
      }
    },
    "/api/v1/businesses/{business_id}": {
      "get": {
        "operationId": "get_business_endpoint_api_v1_businesses__business_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Business Endpoint",
        "tags": [
          "businesses"
        ]
      },
      "patch": {
        "operationId": "update_business_endpoint_api_v1_businesses__business_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Business Endpoint",
        "tags": [
          "businesses"
        ]
      }
    },
...
    "/api/v1/businesses/{business_id}/clients": {
      "get": {
        "operationId": "list_clients_endpoint_api_v1_businesses__business_id__clients_get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClientRead"
                  },
                  "title": "Response List Clients Endpoint Api V1 Businesses  Business Id  Clients Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Clients Endpoint",
        "tags": [
          "clients"
        ]
      },
      "post": {
        "operationId": "create_client_endpoint_api_v1_businesses__business_id__clients_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Client Endpoint",
        "tags": [
          "clients"
        ]
      }
    },
    "/api/v1/businesses/{business_id}/clients/{client_id}": {
      "get": {
        "operationId": "get_client_endpoint_api_v1_businesses__business_id__clients__client_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Client Endpoint",
        "tags": [
          "clients"
        ]
      },
      "patch": {
        "operationId": "update_client_endpoint_api_v1_businesses__business_id__clients__client_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Client Endpoint",
        "tags": [
          "clients"
        ]
      }
    },
    "/api/v1/businesses/{business_id}/clients/{client_id}/bookings": {
      "get": {
        "operationId": "get_client_bookings_endpoint_api_v1_businesses__business_id__clients__client_id__bookings_get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingRead"
                  },
                  "title": "Response Get Client Bookings Endpoint Api V1 Businesses  Business Id  Clients  Client Id  Bookings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Client Bookings Endpoint",
        "tags": [
          "clients"
        ]
      }
    },
...
  },
...
}